# static/openapi_de.yaml
openapi: 3.0.3
info:
  title: TV Channel Editor API
  version: 1.0.0
  summary: API zum Analysieren und Neuverpacken von Smart-TV-Senderlisten
  description: 'Smart-TV-Senderlisten-Dateien analysieren, prüfen, ändern und neu
    verpacken.


    Zu den unterstützten Marken gehören Samsung, LG, Sony und Hisense.

    Die API nutzt einen sitzungsbasierten Arbeitsablauf:


    1. Eine unterstützte TV-Senderlisten-Datei hochladen

    2. Die zurückgegebenen Werte `session_id`, `brand` und das Array `channels` auslesen

    3. Ändern Sie die Kanalobjekte (z. B. Reihenfolge ändern, umbenennen, als Favorit
    markieren/sperren)

    4. Senden Sie die `session_id` und das geänderte `channels`-Array an `/download`

    5. Empfangen Sie die neu erstellte binäre Kanal-Listendatei

    '
  contact:
    name: Turan KAYA (Founder & Security Contact)
    url: https://github.com/tarihcituranx/TVChanneleditor
    email: tarihcituranx@proton.me
servers:
- url: https://tvchanneleditor.onrender.com
  description: Production Server
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: FILE_NOT_FOUND
        error:
          type: string
          example: Dosya bulunamadı
...
    UploadResponse:
      type: object
      required:
      - success
      - brand
      - format
      - session_id
      - channels
      properties:
        success:
          type: boolean
          example: true
        brand:
          type: string
          description: Erkannte TV-Marke.
          enum:
          - samsung
          - lg
          - sony
          - hisense
          - unknown
          example: samsung
        original_filename:
          type: string
          description: Der Name der ursprünglich hochgeladenen Datei; hilfreich zur
            Identifizierung des TV-Modells.
          example: lg_xxLA691S.tll
        format:
          type: string
          description: Internes Dateiformat erkannt.
          enum:
          - scm
          - zip
          - tll
          - sdb.xml
          - servicelist.db
          example: scm
        session_id:
          type: string
          description: Nach erfolgreicher Auswertung wird eine temporäre Sitzungskennung
            zurückgegeben. Übergeben Sie genau diesen Wert an /download.
          example: sess_01J8ABC123XYZ
        channels:
          type: array
          description: Die Kanäle wurden in ihrer aktuellen Reihenfolge analysiert.
          items:
            $ref: '#/components/schemas/Channel'
    DownloadRequest:
      type: object
      required:
      - session_id
      - channels
      properties:
        session_id:
          type: string
          description: Von /upload zurückgegebene Sitzungs-ID.
          example: sess_01J8ABC123XYZ
        channels:
          type: array
          description: Geänderte Kanalliste. Die Reihenfolge im Array bestimmt die
            Reihenfolge der Kanäle im Ergebnis.
          minItems: 1
          items:
            $ref: '#/components/schemas/Channel'
      example:
        session_id: sess_01J8ABC123XYZ
        channels:
        - false: 1
          No: 1

          Slot: 0

          Name: TRT 1 HD

          Type: TV

          Freq: '11958'

          Pol: V

          Sym: '27500'

          Lock: false

          Hide: false

          Skip: false

          Encrypted: 'No'

          Fav1: false

          Fav2: false

          Fav3: false

          Fav4: false

          Fav5: false
security:
- {}
- ApiKeyAuth: []
paths:
  /upload:
    post:
      tags:
      - Channel Lists
      operationId: uploadChannelList
      summary: Eine TV-Senderliste hochladen und auswerten
      description: 'Laden Sie eine Datei mit einer unterstützten TV-Senderliste hoch.
        Der Server erkennt die TV-Marke, analysiert die Sender und erstellt eine temporäre
        Bearbeitungssitzung.


        Unterstützte Formate nach Marke:

        - Samsung: .scm, .zip

        - LG: .tll

        - Sony: sdb.xml

        - Hisense: servicelist.db

        '
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                  description: TV channel list file (.scm, .zip, .tll, .xml, .db)
      responses:
        '200':
          description: Channel list successfully parsed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              examples:
                samsung_upload:
                  summary: Successful Samsung Parse
                  value:
                    success: true
                    brand: samsung
                    format: scm
                    session_id: sess_01J8ABC123XYZ
                    channels:
                    - false: 1
                      No: 1

                      Slot: 0

                      Name: TRT 1 HD

                      Type: TV

                      Freq: '11958'

                      Pol: V

                      Sym: '27500'

                      Lock: false

                      Hide: false

                      Skip: false

                      Encrypted: 'No'

                      Fav1: false

                      Fav2: false

                      Fav3: false

                      Fav4: false

                      Fav5: false
        '400':
          description: Invalid, unsupported, or malformed channel list file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Uploaded file is too large (Max 2MB).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded (50 per hour)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /build:
    post:
      tags:
      - Channel Lists
      operationId: buildChannelList
      summary: Repack modified channel list
      description: 'Rebuild the channel-list file associated with a session. The returned
        JSON contains a `download_url` to fetch the binary file.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
      responses:
        '200':
          description: Rebuild successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  download_url:
                    type: string
                    example: /download/sess_123/channels.scm
        '400':
          description: Invalid request or expired session_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /download/{session_id}/{filename}:
    get:
      tags:
      - Channel Lists
      operationId: downloadBinaryFile
      summary: Download the rebuilt binary file
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
      - name: filename
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Rebuilt binary channel-list file.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/share:
    post:
      tags:
      - Cloud Sync
      operationId: createShareCode
      summary: Create a 8-character share code
      description: Save a channel list draft to the cloud and receive a unique 8-character
        code to share with other devices.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - draft
              properties:
                draft:
                  type: object
                  description: The current channel list and session data
      responses:
        '200':
          description: Share code created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: string
                    example: X8VUJP
                  message:
                    type: string
        '400':
          description: Invalid data
        '429':
          description: Rate limit exceeded (20 per minute)
  /api/share/{code}:
    get:
      tags:
      - Cloud Sync
      operationId: getSharedList
      summary: Retrieve a shared channel list
      description: Fetch a previously saved channel list using its 8-character share
        code.
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
        description: The 8-character share code
      responses:
        '200':
          description: Channel list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '404':
          description: Code invalid or expired
        '429':
          description: Rate limit exceeded (20 per minute)